refactor: phase 3 - move view routes to views blueprint
