+++ src/app.py
@@ -128,33 +128,10 @@ latest_results = {
     'deep_scan': False
 }
 
-@app.route('/', methods=['GET'])
-def landing():
-    return render_template('landing_v2.html')
-
 # @app.route('/landing-v2', methods=['GET'])
 # def landing_v2():
 #     return render_template('landing_v2.html')
 
-@app.route('/dashboard', methods=['GET'])
-def dashboard():
-    return render_template(
-        'dashboard.html',
-        results=latest_results['results'],
-        original_target=latest_results['target'],
-        deep_scan=latest_results['deep_scan'],
-        active_page='dashboard'
-    )
-
-@app.route('/history', methods=['GET'])
-def history_page():
-    history = load_history()
-    print(f"Loading history page. Found {len(history)} items.")
-    return render_template(
-        'history.html',
-        history=history,
-        active_page='history'
-    )
 
 @app.route('/clear', methods=['POST'])
 def clear():
@@ -251,11 +228,6 @@ def database_vulnerability_page():
     
     return render_template('database_vulnerability.html', vulnerabilities=vulnerabilities, message=message, active_page='database-vulnerability')
 
-@app.route('/settings', methods=['GET'])
-def settings_page():
-    """Settings page for scanner configuration"""
-    return render_template('settings.html', active_page='settings')
-
 @app.route('/api/save-settings', methods=['POST'])
 def save_settings_api():
     """API endpoint to save scanner settings to session"""
@@ -280,30 +252,6 @@ def get_settings_api():
         logger.error(f"Error retrieving settings: {e}")
         return jsonify({'status': 'error', 'message': str(e)}), 500
 
-@app.route('/contact', methods=['GET', 'POST'])
-def contact():
-    if request.method == 'POST':
-        name = request.form.get('name')
-        email = request.form.get('email')
-        subject = request.form.get('subject')
-        message = request.form.get('message')
-
-        if not all([name, email, subject, message]):
-            return render_template('contact.html', error="All fields are required.", active_page='contact')
-
-        message_data = {
-            'name': name,
-            'email': email,
-            'subject': subject,
-            'message': message
-        }
-
-        if save_message(message_data):
-            return render_template('contact.html', success="Your message has been sent successfully!", active_page='contact')
-        else:
-            return render_template('contact.html', error="There was an error sending your message. Please try again later.", active_page='contact')
-
-    return render_template('contact.html', active_page='contact')
 
 @app.route('/export/<scan_id>', methods=['GET'])
 def export_report(scan_id):
